telegram-bot-service/pyproject.toml
[tool.poetry]
name = "poetry init"
version = "0.1.0"
description = ""
authors = ["DarthZiel <Studerzhan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.7.0"
python-dotenv = "^1.0.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.30"}
asyncpg = "^0.29.0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
